analysis/requirements.txt
streamlit
numpy
pandas
altair
pillow
opencv-python-headless
